Updated sbt Dockerfile to handle diffrent global sbtopts location
As a part of the sbt's docker build, the global sbtopts is modified
("-Dsbt.override.build.repos=true" is appended to it). However, the
location of the sbtopts is different amongst different sbt versions.

At the time of this commit, there are only two possible locations of the
sbtopts across 29 different versions of sbt. The sbt Dockerfile checks
the existence of the sbtopts in these two locations before appending to
it and fails if it does not exist in one of these locations