Add scala and sbt versions needed for Gatling tests (#6)
* add sane labels to Dockerfiles

* Add scala and sbt versions needed for Gatling tests

These versions of scala and sbt are needed by the gatling plugin used by
the performance test project.

* Updated sbt Dockerfile to handle diffrent global sbtopts location

As a part of the sbt's docker build, the global sbtopts is modified
("-Dsbt.override.build.repos=true" is appended to it). However, the
location of the sbtopts is different amongst different sbt versions.

At the time of this commit, there are only two possible locations of the
sbtopts across 29 different versions of sbt. The sbt Dockerfile checks
the existence of the sbtopts in these two locations before appending to
it and fails if it does not exist in one of these locations

* add test for scala and sbt new version

* fix changing java_home version