Added Python versions 3.4 and 3.6
Note that for Python versions >=3.4 pip comes installed out of the box.
Both of these versions are built from source and the location of the
compiled binaries is:
/usr/local/bin/python3.x
/usr/local/bin/pip3.x

Note that the version of pip installed when python is built from source
is outdated as of this commit. If this is a problem, the Dockerfile
shall be amended to update the pip version to the latest.

This PR also makes whitespace changes and minor modifications to address
issues flagged by yamllint