Revert "Jenkins slave R 3.5.0-1 dockerfile (#5)" (#12)
This reverts commit 84911219bbdc1dc12a736e11ad93bf22c5313ae5.

It was thought that the team which needs R should shoulder the
responsibility of creating a docker image for it. The commit being
reverted (hash 8491121) can serve as the basis for this.

At the time of this revert, the R docker build fails on compiling
'epel-release' and blocks any other changes to the repository.