Amend the python build to create executable linked to shared library
This commit fixes the issue that the python executable built requires
specifiying the LD_LIBRARY_PATH to load the correct library versions.

Replacing `altinstall` with `install` ensures that the python binary is
built in the default prefix (/usr/local/bin) and since that has
precedence on the PATH, this allows existing Jenkinsfile to use
`python` without specifying the full path and still get the expected
version of python.

Currently, with altinstall, a python executable is not built in
/usr/local/bin and `python` then defaults to /usr/bin or /bin (which are
the system versions, probably not the same as the one expected from the
image)

Additioanlly I have begun using install instead of altinstall so that
the python executable on the

This issue is seen by comparing the output of `python -V` with
`LD_LIBRARY_PATH=/usr/local/lib python -V` . The former returns an
incorrect version (2.7.5) whereas the latter return the correct version
(2.7.15).

The default prefix when building python from source is /usr/local.
However the executable built loads the shared libraries from /lib64
(symlinked to /usr/lib64) which ends up loading the (incorrect) system
libpython2.7.so instead of loading it from the one that is built in
/usr/local/lib

The libraries being loaded by the executable are visible by executing:
For the correctly built exectuable it is :
```
[root@90fbc9dd6cfa /]# ldd /usr/local/bin/python
	libpython2.7.so.1.0 => /usr/local/lib/libpython2.7.so.1.0 (0x00007fedc6348000)
```
whereas for the current version (before this commit) the so is loaded
from /lib64 (-> /usr/lib64)

The fix is to either:
1. Specify the library path like so before invoking the python
exectuable:
  `LD_LIBRARY_PATH=/usr/local/lib /usr/local/bin/python -V`

OR,

2. Specify the library path at compile time using LD_RUN_PATH (as done
in this commit) so that it does not need to be specified at runtime and
does not require alterations to existing Jenkinsfile

Relevant links:
http://blog.dscpl.com.au/2015/06/installing-custom-python-version-into.html
https://bugs.python.org/issue27685