Complete rewrite of makefile to remove redundancy and duplication
This commit reimplements the targets as simple calls to a makefile
'meta-function', which dynamically sets 3 variables and used them to run
the commands to check individual containers.

This commit makes it very simple to add support for future tool
versions. For example, adding support to a new scala version would
simply require amending the SCALA_VERSION_LIST variable.